feat: parseDate function
- 이름이 2개가 겹쳐서 파일 분리하고 네이밍 변경
