+++ src/app/sitemap.ts
@@ -1,15 +1,6 @@
 import { getAllPost } from "@/lib/api/notion"
+import { localDateParseToNumberDate } from "@/lib/utils/date"
 import { MetadataRoute } from "next"
-
-/**
- * yyyy년 mm월 dd일 형식의 문자열을 Date 객체로 변환합니다.
- */
-function parseDate(dateString: string) {
-  const dateParts = dateString.match(/(\d{4})년\s*(\d{1,2})월\s*(\d{1,2})일/)
-  // eslint-disable-next-line
-  const [_, year, month, day] = dateParts || []
-  return new Date(Number(year), Number(month) - 1, Number(day))
-}
 
 export default async function sitemap({
   id,
@@ -21,7 +12,7 @@ export default async function sitemap({
   const articles = await getAllPost()
   const url = articles.map((article) => ({
     url: `${BASE_URL}/articles/${article.slug}`,
-    lastModified: parseDate(article.createdTime).toISOString(),
+    lastModified: localDateParseToNumberDate(article.createdTime).toISOString(),
   }))
 
   return [
