+++ src/lib/utils/utils.ts
@@ -2,14 +2,6 @@ export function generateSlug(title: string) {
   return encodeURIComponent(title.toLowerCase().replace(/ /g, "-"))
 }
 
-export function parseDate(date: string) {
-  return new Date(date).toLocaleDateString("ko-KR", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  })
-}
-
 export function onClickGithub() {
   window.open("https://github.com/obvoso", "_blank")
 }
